fix(dashboard): モック集計の整合とページ metadata を修正
部署未選択時のモック値を全部門扱いで揃え、client 化で失われていたダッシュボードのタイトル設定も server wrapper で復元する。
