+++ frontend/src/lib/dashboard/mockData.ts
@@ -1,6 +1,7 @@
 // フィルター入力からダッシュボード表示データを合成するモックロジック。
 // 本実装では集計バックエンド（人事DB / ポイント DB）と連携する想定 /* TODO: API 連携 */。
 
+import { HD_DEPARTMENTS, SAIBU_HQ } from "./org";
 import type { DashboardData, DashboardFilter } from "./types";
 
 const clamp = (v: number, lo: number, hi: number) => Math.max(lo, Math.min(hi, v));
@@ -8,8 +9,17 @@ const clamp = (v: number, lo: number, hi: number) => Math.max(lo, Math.min(hi, v
 const ANNUAL_TARGET = 6000;
 
 export function computeDashboardData(filter: DashboardFilter): DashboardData {
+  const availableDepartments =
+    filter.company === "HD"
+      ? HD_DEPARTMENTS.length
+      : filter.hq
+        ? SAIBU_HQ[filter.hq].departments.length
+        : Object.values(SAIBU_HQ).reduce((sum, hq) => sum + hq.departments.length, 0);
+  // 部署未選択は「全部門」扱いで、比較表示と数値ロジックの意味を揃える。
+  const selectedDepartments =
+    filter.departments.length === 0 ? availableDepartments : filter.departments.length;
   // 部署数・役職数で揺らす簡易係数（範囲に丸める）
-  const deptFactor = clamp(filter.departments.length / 3, 0.3, 1.5);
+  const deptFactor = clamp(selectedDepartments / 3, 0.3, 1.5);
   const roleFactor = clamp(filter.roles.length / 3, 0.3, 1.2);
   const f = deptFactor * roleFactor;
